Fix bug with gradient leaks at the GAN example.
Following #591, the GAN example at the Pytorch Lightning project is leaking
gradients between the Generator and the Discriminator steps. This
fix clears the gradients of the Discriminator before returning its loss.
In the future, fix #603 may deem this fix unnecessary.

This diff also adds a note for using SyncBatchNorm in case of training with DDP.