Updates theme Sphinx configuration
I am updating the project's Sphinx documentation to fix (#819). The issue is related to a library the Sphinx extension `nbsphinx` (to load Jupyter Notebooks) loads into the docs context (RequireJS). That library conflicts with other theme libraries, causing the latter to be not loaded. This would result in several crashes, the most obvious of them the lack of anchors.

The fix above solves all errors -- and now anchors work.