Fixed example implementation of AutoEncoder.
The previous implementation trained a auto encoder and evaluated
classificator.
I try to fix this by replacing the evaluation metric with an auto
encoder metric.
Hence, no classification is done.
I'm not 100% sure what the original authors intent was, since he
extends a classification model (LitMNIST) but does not use it.
The following model is an AutoEncoder and does not do any
classification.

 1. Small textual changes.
 2. forward() now implements encoding and not decoding (as it was described
 in the text.)
 3. _shared_eval uses MSE loss instead of class loss, since no
 classification weights are learned.
 4. initialized MSE in __init__, since calling MSE directly is not
 supported.