Initialize process subgroup in DDP plugin instead of post-localSGD comm hook to address the flakiness
When `ddp_comm_state` is provided in `DDPPlugin` ctor, if the state is `PostLocalSGDState`, no initialized process subgroup can be provided at that time, because the initialization of a subgroup is on the basis of the global process group, which is initialized after `DDPPlugin` ctor. Therefore, previously the subgroup is initialized in the communication hook (in PyTorch Distributed codebase). However, it turns out that such subgroup initialization can somehow cause flakiness.

Therefore, now move subgroup initialization out of communication hook, and now it is the caller's responsibility to ensure the correctness of all the args.