Typo in major heading seen by newcomers
Correct typo in one of the first major headings newcomers to Lightning see when they are considering migrating their code to use Lightning.

I know this is a trivial change in terms of the text change itself, but I really think it's valuable for one of the most important landing pages that users first investigating Lightning see - to have rock-solid, professional text without obvious typos. Here was a typo in the main heading itself. I suggest fixing it straightaway via this PR.