Support Injecting Secrets into Apps Running in the Cloud (#14612)
Adds a new '--secret' flag to 'lightning run app':

lightning run app --cloud --secret MY_SECRET=my-secret-name app.py

When the Lightning App runs in the cloud, the 'MY_SECRET'
environment variable will be populated with the value of the
referenced Secret. The value of the Secret is encrypted in the
database, and will only be decrypted and accessible to the
Flow/Work processes in the cloud.

Co-authored-by: Sherin Thomas <sherin@grid.ai>
Co-authored-by: Noha Alon <nohalon@gmail.com>
Co-authored-by: thomas chaton <thomas@grid.ai>

(cherry picked from commit 71719b9960403322cdb43aa2dc13a2e62d840be9)