Prevent bug when launching apps on multiple clusters (#15226)
Stops a bug when cross-launching an app between clusters.

Currently the platform does not allow running multiple app instances. If you have `app-1` running on `cluster-1` and try to run it on `cluster-2`, the CLI will succeed but the app will never start.

This PR prevents this disconnect. The app should not be uploaded / released if it won't run. An error is presented to the user explaining what happened and how to proceed (specify a different `--name`: e.g. `app-2`).

Once the platform supports multiple app instances / running individual apps on multiple clusters, this PR can be reverted.