Updated trainer to not call self.trainer.model during validation.
Before, using the trainer would result in:

AttributeError: Your LightningModule code tried to access `self.trainer.model` but this attribute is not available when using Fabric with a LightningModule.

Which would occur during validation when 'on_validation_model_eval' and 'on_validation_model_train' were called, as they do not work with fabric. This change allows the model to be switched between train and eval mode (which the original behavior intends) without relying on calling 'self.trainer.model'. A check is added for overridden methods to not change a user's code.