Move checkpointing out of signal handlers
Signal handlers are a precarious environment, because they can be called at any time, for
example in the middle of another function. So performing IO in there can lead to difficult
to track down errors. This change moves the heavy lifting of requeuing into the normal
life cycle of lightning loops.